refactor: extract shared form helpers for validation and dirty state management
