+++ src/pages/admin/who-we-are/components/modals/forms/translate-description-form/TranslateWhoWeAreDescriptionForm.tsx
@@ -1,12 +1,17 @@
 import { useFormManager } from '@/hooks/admin/use-form-manager/useFormManager';
 import { WHO_WE_ARE_VALIDATION_FUNCTIONS } from '@/validation/admin/who-we-are-schema/WhoWeAreSchema';
-import { forwardRef, useEffect, useRef } from 'react';
+import { forwardRef } from 'react';
 import styles from './TranslationWhoWeAreDescriptionForm.module.scss';
 import cn from 'classnames';
 import { WHO_WE_ARE_TEXT } from '@/const/admin/who-we-are';
 import { RichTextInputGroup } from '@/components/admin/input-groups/rich-text-input-group/RichTextInputGroup';
 import { GeneralFormProps, GeneralFormRef } from '../../strategies/who-we-are-modal-strategy';
 import { getPlainTextFromHtml } from '@/utils/functions/get-plain-text-from-html/get-plain-text-from-html';
+import {
+    getWhoWeAreTextValidationError,
+    useDirtyChangeEffect,
+    useReadyRef,
+} from '@/pages/admin/who-we-are/components/modals/forms/shared/translate-who-we-are-form-helpers';
 
 export interface TranslateWhoWeAreDescriptionFormValues {
     description: string;
@@ -65,23 +70,12 @@ export const TranslateWhoWeAreDescriptionForm = forwardRef<
             onSubmit: (data, _status) => onSubmit(data),
         });
 
-        const isReadyRef = useRef(false);
+        const isReadyRef = useReadyRef();
 
-        useEffect(() => {
-            const id = setTimeout(() => {
-                isReadyRef.current = true;
-            }, 0);
-            return () => clearTimeout(id);
-        }, []);
-
-        useEffect(() => {
-            const isDirty = JSON.stringify(formState) !== JSON.stringify(initialData);
-            onDirtyChange?.(isDirty);
-        }, [formState, initialData, onDirtyChange]);
+        useDirtyChangeEffect(formState, initialData, onDirtyChange);
 
         const validateAndSetFieldError = (field: FormFieldName, value: string) => {
-            const plainText = getPlainTextFromHtml(value).trim();
-            const error = WHO_WE_ARE_VALIDATION_FUNCTIONS.validateText(plainText);
+            const error = getWhoWeAreTextValidationError(value);
 
             setErrors((prev) => ({ ...prev, [field]: error }));
         };
