+++ src/pages/admin/who-we-are/components/modals/forms/translate-multiple-descriptions-form/TranslateWhoWeAreMultipleDescriptionsForm.tsx
@@ -1,12 +1,17 @@
 import { useFormManager } from '@/hooks/admin/use-form-manager/useFormManager';
 import { WHO_WE_ARE_VALIDATION_FUNCTIONS } from '@/validation/admin/who-we-are-schema/WhoWeAreSchema';
-import { forwardRef, useEffect, useRef } from 'react';
+import { forwardRef, useRef } from 'react';
 import styles from './TranslateWhoWeAreMultipleDescriptionsForm.module.scss';
 import cn from 'classnames';
 import { WHO_WE_ARE_TEXT } from '@/const/admin/who-we-are';
 import { RichTextInputGroup } from '@/components/admin/input-groups/rich-text-input-group/RichTextInputGroup';
 import { GeneralFormProps, GeneralFormRef } from '../../strategies/who-we-are-modal-strategy';
 import { getPlainTextFromHtml } from '@/utils/functions/get-plain-text-from-html/get-plain-text-from-html';
+import {
+    getWhoWeAreTextValidationError,
+    useDirtyChangeEffect,
+    useReadyRef,
+} from '@/pages/admin/who-we-are/components/modals/forms/shared/translate-who-we-are-form-helpers';
 
 interface TranslateDescriptionRow {
     contentId: number;
@@ -71,7 +76,7 @@ export const TranslateWhoWeAreMultipleDescriptionsForm = forwardRef<
             rows: buildDefaultRows(initialData),
         });
 
-        const isReadyRef = useRef(false);
+        const isReadyRef = useReadyRef();
         const touchedRowsRef = useRef(new Set<number>());
 
         const { formState, setFormState, errors, setErrors, isSubmitting } = useFormManager<
@@ -86,21 +91,10 @@ export const TranslateWhoWeAreMultipleDescriptionsForm = forwardRef<
             onSubmit: (data, _status) => onSubmit(data),
         });
 
-        useEffect(() => {
-            const id = setTimeout(() => {
-                isReadyRef.current = true;
-            }, 0);
-            return () => clearTimeout(id);
-        }, []);
-
-        useEffect(() => {
-            const isDirty = JSON.stringify(formState) !== JSON.stringify(initialData);
-            onDirtyChange?.(isDirty);
-        }, [formState, initialData, onDirtyChange]);
+        useDirtyChangeEffect(formState, initialData, onDirtyChange);
 
         const validateAndSetRowDescriptionError = (rowIndex: number, value: string) => {
-            const plainText = getPlainTextFromHtml(value).trim();
-            const error = WHO_WE_ARE_VALIDATION_FUNCTIONS.validateText(plainText);
+            const error = getWhoWeAreTextValidationError(value);
 
             setErrors((prev) => {
                 const nextRows = [...(prev.rows ?? [])];
